restructure and check for CrossCorrelations udf when UI is launched
