+++ seeq/addons/correlation/__init__.py
@@ -1,5 +1,5 @@
 from ._version import __version__
-from ._seeq_formulas import correlation_udfs, signals_from_formula
+from ._seeq_formulas import correlation_udfs, signals_from_formula, create_udfs
 from ._preprocessor import _validate_df, default_preprocessing_wrapper
 from ._cross_correlations import cross_corr_matrix_raw, cross_corr_matrix_lagged, lags_coeffs
 from ._heatmap import _heatmap_plot, heatmap
@@ -13,4 +13,4 @@ __all__ = ['__version__', 'cross_corr_matrix_raw', 'cross_corr_matrix_lagged', '
            '_heatmap', 'create_worksheet', '_create_worksheet', 'get_workbook', 'worksheet_with_lagged_signals',
            'create_lagged_signals', 'pairplot', 'worksheet_corrs_and_time_shifts',
            'correlation_udfs', 'signals_from_formula', 'CorrelationHeatmap', '_validate_df', 'lags_coeffs',
-           'default_preprocessing_wrapper', 'utils']
+           'default_preprocessing_wrapper', 'utils', 'create_udfs']
