+++ seeq/addons/correlation/utils/__init__.py
@@ -1,10 +1,12 @@
 from ._common import validate_argument_types, print_red, path_list_to_string, path_string_to_list
 from ._cache_management import clear_cache_all
-from ._permissions import get_user, get_user_group
-from ._sdl import pull_only_signals, get_worksheet_url, get_workbook_worksheet_workstep_ids, get_seeq_url
+from ._permissions import get_user, get_user_group, DEFAULT_USERS, DEFAULT_GROUP
+from ._sdl import (pull_only_signals, get_worksheet_url, get_workbook_worksheet_workstep_ids, get_seeq_url,
+                   check_udf_package)
 from ._seeq_new_content import create_condition, create_workstep_signals
 
 
 __all__ = ['validate_argument_types', 'print_red', 'create_condition', 'create_workstep_signals', 'get_user',
            'get_user_group', 'pull_only_signals', 'get_worksheet_url', 'get_workbook_worksheet_workstep_ids',
-           'clear_cache_all', 'get_seeq_url', 'path_list_to_string', 'path_string_to_list']
+           'clear_cache_all', 'get_seeq_url', 'path_list_to_string', 'path_string_to_list', 'check_udf_package',
+           'DEFAULT_USERS', 'DEFAULT_GROUP']
